refs #472 Removed static_cast<int> for enums in blackcore, no longer needed.

+++ src/blackgui/components/enableforruntime.cpp
@@ -118,7 +118,7 @@ namespace BlackGui
         void CEnableForRuntime::playNotifcationSound(BlackSound::CNotificationSounds::Notification notification) const
         {
             if (!this->getIContextAudio()) return;
-            this->getIContextAudio()->playNotification(static_cast<uint>(notification), true);
+            this->getIContextAudio()->playNotification(notification, true);
         }
     }
 }
